InputControllers Commit 2 (not working)
InputController explanation:
I’ve created it very similarly to the AILab. The InputController is an
interface that contains the east,north, etc. movement fields and calls
getAction.
PlayerController returns the action of the ship but does not actually
process it. As the game progresses this will be done in
CollisionController probably but for now the actually movement update
is in GameMode. PlayerController also has the getClickPos that will
give the coordinates of the mouse.

In Hunter, the update method takes in a controlCode and updates the
hunter’s position (velocity and angle) accordingly.

Finally in GameMode, updateGame() gets the action from the
PlayerController. Hunter then updates with that action, then with the
new velocity given it actually changes the hunter’s position.